Move to helper widgets, fix typo, refactor code
