+++ spyderlib/widgets/helperwidgets.py
@@ -8,15 +8,18 @@
 Helper widgets.
 """
 
+# Third party imports
 from spyderlib.qt.QtCore import QPoint, QSize, Qt
 from spyderlib.qt.QtGui import (QToolButton, QToolTip,
                                 QStyledItemDelegate, QApplication,
                                 QTextDocument, QStyleOptionViewItem,
                                 QAbstractTextDocumentLayout, QStyle,
-                                QVBoxLayout, QSpacerItem,
-                                QMessageBox, QCheckBox)
+                                QVBoxLayout, QSpacerItem, QPainter,
+                                QMessageBox, QCheckBox, QLineEdit)
 
+# Local imports
 from spyderlib.config.base import _
+from spyderlib.utils import icon_manager as ima
 from spyderlib.utils.qthelpers import get_std_icon
 
 
@@ -138,6 +141,71 @@ class HTMLDelegate(QStyledItemDelegate):
         doc.setTextWidth(options.rect.width())
 
         return QSize(doc.idealWidth(), doc.size().height())
+
+
+class IconLineEdit(QLineEdit):
+    """Custom QLineEdit that includes an icon representing the validation."""
+    def __init__(self, *args, **kwargs):
+        super(IconLineEdit, self).__init__(*args, **kwargs)
+
+        self._status = True
+        self._status_set = True
+        self._valid_icon = ima.icon('todo')
+        self._invalid_icon = ima.icon('warning')
+        self._set_icon = ima.icon('todo_list')
+
+    def _refresh(self):
+        """After an application style change, the paintEvent updates the
+        custom defined stylesheet.
+        """
+        application_style = QApplication.style().objectName()
+        padding = self.height()
+        css_base = """QLineEdit {{border: none;
+                                 padding-left: {padding}px;
+                                 }}
+                   """
+        css_oxygen = """QLineEdit {{background: transparent;
+                                   border: none;
+                                   padding-left: {padding}px;
+                                   }}
+                     """
+        if application_style == 'oxygen':
+            css_template = css_oxygen
+        else:
+            css_template = css_base
+
+        css = css_template.format(padding=padding)
+        self.setStyleSheet(css)
+
+    def update_status(self, value, value_set):
+        """Update the status and set_status to update the icons to display."""
+        self._status = value
+        self._status_set = value_set
+        self.repaint()
+        self.update()
+
+    def paintEvent(self, event):
+        """Qt Override.
+
+        Include a validation icon to the left of the line edit.
+        """
+        self._refresh()
+        super(IconLineEdit, self).paintEvent(event)
+        painter = QPainter(self)
+
+        rect = self.geometry()
+        space = int((rect.height())/6)
+        h = rect.height() - space
+
+        if self._status and self._status_set:
+            pixmap = self._set_icon.pixmap(h, h)
+        elif self._status:
+            pixmap = self._valid_icon.pixmap(h, h)
+        else:
+            pixmap = self._invalid_icon.pixmap(h, h)
+
+        painter.drawPixmap(space, space, pixmap)
+
 
 
 def test_msgcheckbox():
